Added multiple products selection
Multiple products can now be added to an invoice (in the frontend. I'm used Alpine JS to create a template for the products dropdown and counter for product amount. When the user wants to add another product, they click the button and the template is cloned and added to the table.